Drastically reduce cloud pin stripping
Not an ideal solution, but a possible good interim one.

The low cloud pin stripping appears to be roughly linearly tied to the height of the low cloud layer (lower the cloud layer, issue is much worse. Raise the clouds, issue drastically reduced)

By increasing the height of the lower cloud layer to 4.2 km the visual impact is greatly reduced.